pylinting and restructuring
