+++ .travis.yml
@@ -12,4 +12,4 @@ python:
 
 install: "pip install -r requirements.txt"
 
-script: python -m unittest test_randomfact.py
+script: cd tests && python3 -m unittest test_randomfact.py
