+++ README.md
@@ -4,3 +4,14 @@ This is the first thing I wrote with Python, unit tests, and CI!
 It's very simple - just a Flask app that shows some fun stuff about Pandas.
 
 I'm looking back at this two years later and seeing what I've learned since then (and making some updates based on the learnings :) ).
+
+
+# Currently working on
+Linting python files (3 test files left!) and organizing the doc tree.
+
+# For the future
+Improve code quality.
+Review CD.
+How to deal with secrets (i.e. terraform aws keys).
+
+Add more fun stuff about pandas to explore flask further!
