+++ data/pandafacts.txt
@@ -1,0 +1,7 @@
+Pandas are usually quiet and solitary animals, but they may use a variety of sounds to communicate to each other. They may squeak, growl, bark and huff. Panda researchers have counted 11 different panda calls - and four of them are used only when searching for a mate.
+
+Pandas eat almost nothing but bamboo shoots and leaves. Occasionally they eat other vegetation, fish, or small animals, but bamboo accounts for 99 percent of their diets. Pandas eat fast, they eat a lot, and they spend about 12 hours a day doing it. The reason: They digest only about a fifth of what they eat. Overall, bamboo is not very nutritious. To stay healthy, they have to eat a lot - up to 15 percent of their body weight in 12 hours - so they eat fast.
+
+A panda mother is 900 times bigger than her newborn cub, which weighs less than 5 oz. In comparison, an 8-pound human baby would have a mother that weighed 7,200 pounds.
+
+A giant panda's face is cute, but it is not chubby. It gets its shape from massive cheek muscles. In fact, a giant panda's jaw and cheek muscles are so powerful that a panda can easily chew an aluminum dish into tiny pieces. They can also easily bite through a thick bamboo stalk. Humans have trouble cutting the same stalk with an ax.
