--- functional_tests.py
@@ -1,6 +1,0 @@
-from selenium import webdriver
-
-browser = webdriver.Firefox()
-browser.get('http://localhost:5000')
-
-assert 'Flask' in browser.title
