+++ requirements.txt
@@ -1,2 +1,3 @@
 Flask==1.1.1
 pylint==2.3.1
+selenium==3.141.0
