+++ templates/index.html
@@ -1,0 +1,9 @@
+
+<!doctype html>
+<html>
+<body>
+<p> Hello world!</p>
+<p><a href="{{ url_for('pandafacts') }}"> what do you know about pandas?</a></p>
+<p><a href="{{ url_for('pandagifs') }}"> look at this panda!</a></p>
+</body>
+</html>
